Handle starting in help window
If the current window was the open help window chosen by tagurl, then
returning to the previous buffer would bring you somewhere other than
where you started.

This fixes that, correctly returning the cursor to the original
file, but in the same window.

Additionally, handles the case where the queried tag is in the same help
file, just moving back to the original position.